requirements.txt
python3-psycopg2
python3-pylint
python3-pytest
python3-pytest-cov